+++ src/helpers/listener.js
@@ -1,8 +1,6 @@
 import { resolve } from 'path';
 import { getCurrentDirectory } from './console.js';
 import { unlink } from 'fs/promises';
-import { createReadStream, createWriteStream } from 'fs';
-import { pipeline } from 'stream/promises';
 
 import { up, cd } from '../operations/navigation.js';
 import { ls } from '../operations/ls.js';
@@ -10,23 +8,7 @@ import { cat } from '../operations/cat.js';
 import { add } from '../operations/add.js';
 import { rn } from '../operations/rn.js';
 import { cp } from '../operations/cp.js';
-
-const mv = async (path, newPath) => {
-  try {
-    if (path && newPath) {
-      const currentDir = process.cwd();
-      const readStream = createReadStream(resolve(currentDir, path), {encoding: 'utf-8'});
-      const writeStream = createWriteStream(resolve(currentDir, newPath), {encoding: 'utf-8'});
-      await pipeline(readStream, writeStream);
-      await unlink(resolve(currentDir, path));
-      getCurrentDirectory();
-    } else {
-      console.log('Invalid input');
-    }
-  } catch {
-    console.log('Operation failed');
-  }
-}
+import { mv } from '../operations/mv.js';
 
 const rm = async (path) => {
   try {
